requirements.txt
Flask
gTTS
requests
python-dotenv
